Initial implementation of DHT22

+++ components/light/include/light.h
@@ -1,5 +1,7 @@
 #ifndef _LIGHT_H_
 #define _LIGHT_H_
+#include <stdint.h>
+
 #include "relay.h"
 #include "iot.h"
 
@@ -11,7 +13,7 @@ typedef struct {
     char name[7];
 }Light_t;
 
-void lightInit(Light_t *light, int pin);
+void lightInit(Light_t *light, int8_t pin);
 void lightToggle(Light_t *light);
 void lightSetState(Light_t *light, RelayState state);
 void lightSendUpdate(Light_t *light, MQTTClient *client, const char *topic);
