+++ components/relay/include/relay.h
@@ -7,11 +7,11 @@ typedef enum RelayState {
 } RelayState; 
 
 typedef struct Relay {
-    int pin;
+    int8_t pin;
     RelayState state;
 } Relay_t;
 
-void relayInit(int pin, Relay_t *relay);
+void relayInit(int8_t pin, Relay_t *relay);
 void relaySetState(Relay_t *relay, RelayState state);
 RelayState relayGetState(Relay_t *relay);
